switching to CommonJS
