+++ src/check-wifi.js
@@ -1,4 +1,4 @@
-import wifi from 'node-wifi';
+const wifi = require('node-wifi');
 
 // Initialize wifi module
 wifi.init({
@@ -7,6 +7,10 @@ wifi.init({
 
 const checkWifiConnection = async () => {
   try {
+    // Ensure wifi is initialized before using other methods
+    const networks = await wifi.scan();
+    // console.log(networks); // Example: log scanned networks
+
     const currentConnections = await wifi.getCurrentConnections();
     if (currentConnections.length > 0) {
       return true;
@@ -14,8 +18,9 @@ const checkWifiConnection = async () => {
       return false;
     }
   } catch (error) {
+    console.error('Error checking wifi connection:', error);
     return false;
   }
 };
 
-export default checkWifiConnection;
+module.exports = checkWifiConnection;
